+++ tests/fixtures/modules/terraform-azure-avm-res-mock/examples/default/main.tf
@@ -4,6 +4,7 @@ module "test" {
   location                 = "westus3"
   create_example_resources = false
   enable_telemetry         = var.enable_telemetry
+  telemetry_location       = var.telemetry_location
 }
 
 output "resource_ids" {