+++ tests/fixtures/modules/terraform-azure-avm-res-mock/examples/ignored_example/main.tf
@@ -1,6 +1,7 @@
 module "test" {
   source = "../../"
 
-  location         = "westus3"
-  enable_telemetry = var.enable_telemetry
+  location           = "westus3"
+  enable_telemetry   = var.enable_telemetry
+  telemetry_location = var.telemetry_location
 }