+++ tests/fixtures/modules/terraform-azure-avm-res-mock/examples/second_example/main.tf
@@ -4,6 +4,7 @@ module "test" {
   location              = "westus3"
   create_mock_resources = false
   enable_telemetry      = var.enable_telemetry
+  telemetry_location    = var.telemetry_location
 }
 
 output "example_resource_ids" {