+++ tests/fixtures/modules/terraform-azurerm-avm-res-mock/examples/default/main.tf
@@ -5,6 +5,7 @@ provider "azurerm" {
 module "test" {
   source = "../../"
 
-  location         = "westus3"
-  enable_telemetry = var.enable_telemetry
+  location           = "westus3"
+  enable_telemetry   = var.enable_telemetry
+  telemetry_location = var.telemetry_location
 }